test: refactor and ci fix for other tests
