+++ package.json
@@ -20,7 +20,8 @@
     "doc-fix": "run-s markdown-lint-fix prettier-fix",
     "doc-fix-everything": "run-s doc-fix doc-fence-check lint-documentation",
     "doc-fence-check": "node tools/check-fenced-code.mjs",
-    "lint-documentation": "jest --coverage false test/documentation.spec.ts",
+    "lint-documentation": "jest --coverage false test/docs/**.spec.ts",
+    "lint-other": "jest --coverage false test/other/**.spec.ts",
     "eslint": "eslint . --cache --cache-location .cache/eslint --report-unused-disable-directives",
     "eslint-fix": "eslint --cache --cache-location .cache/eslint --fix . --report-unused-disable-directives",
     "eslint-ci": "eslint . --cache --cache-strategy content --cache-location .cache/eslint --format gha",
@@ -53,7 +54,7 @@
     "test-e2e:run": "cd test/e2e && npm test",
     "test-schema": "run-s create-json-schema",
     "test:docs": "node --test tools/docs/test/**/*.mjs",
-    "schedule-test-shards": "SCHEDULE_TEST_SHARDS=true ts-node jest.config.ts",
+    "schedule-test-shards": "SCHEDULE_TEST_SHARDS=true ts-node tools/jest-shards.ts",
     "tsc": "tsc",
     "type-check": "run-s 'generate:*' 'tsc --noEmit {@}' --",
     "update-static-data": "run-s 'update-static-data:*'",
